+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-type-checkbox/src/com/liferay/dynamic/data/mapping/type/checkbox/CheckboxDDMFormFieldRenderer.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.type;
+package com.liferay.dynamic.data.mapping.type.checkbox;
 
 import com.liferay.portal.kernel.template.Template;
 import com.liferay.portal.kernel.template.TemplateConstants;
