+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-type-checkbox/src/com/liferay/dynamic/data/mapping/type/checkbox/CheckboxDDMFormFieldType.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.type;
+package com.liferay.dynamic.data.mapping.type.checkbox;
 
 import com.liferay.portlet.dynamicdatamapping.registry.DDMFormFieldRenderer;
 import com.liferay.portlet.dynamicdatamapping.registry.DDMFormFieldType;
