+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-type-checkbox/test/unit/com/liferay/dynamic/data/mapping/type/checkbox/CheckboxDDMFormFieldValueRendererAccessorTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.type;
+package com.liferay.dynamic.data.mapping.type.checkbox;
 
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.language.LanguageUtil;
