+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-type-text/src/com/liferay/dynamic/data/mapping/type/text/TextDDMFormFieldValueAccessor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.type;
+package com.liferay.dynamic.data.mapping.type.text;
 
 import com.liferay.portlet.dynamicdatamapping.model.Value;
 import com.liferay.portlet.dynamicdatamapping.registry.DDMFormFieldValueAccessor;
