+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-type-text/test/unit/com/liferay/dynamic/data/mapping/type/text/TextDDMFormFieldValueRendererAccessorTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.type;
+package com.liferay.dynamic.data.mapping.type.text;
 
 import com.liferay.portal.kernel.util.Html;
 import com.liferay.portal.kernel.util.HtmlUtil;
